Cache Callable.multiResolve
